requirements.txt
yfinance==0.2.51
requests==2.32.3
python-dotenv==1.0.1
rapidfuzz==3.10.1
